fix(assets): change how assets are linked
The `path.join` is designed to be used with n arguments and then puts them together
with the appropriate path separator for the operating system. Even though this will
only ever run on Linux, we should use the functions how they're designed

This commit isn't much of an issue. It really exists to trigger a new release of the
form-web-app after all the problems we've had with getting it working.