fix(mongodb): improve the mongodb connection in cosmos
The packages now take a mongodb url (with or without the database name) and allow the
db name to be configured separately. For the mongoose applications, there is a also an
optional auto index flag.

The FWA mongodb session includes a patch for a
[PR I've raised](https://github.com/mongodb-js/connect-mongodb-session/pull/91) on the
library as the expires session does not fully support Cosmos. It uses a TTL index, but
specifies it on a column called `expires` - Cosmos requires that they're called `_ts`.
There is also a slight divergence with MongoDB in that MongoDB starts the timer when the
date in the column is reached whereas Cosmos starts the timer when the record is added.

If/when this PR is accepted by the library maintainer, will need to update this project with
the new version.