fix(pdf-service-api): remove healthchecks from pdf-service-api
We need to run the PDF Service API container differently in the local dev env from the other
services due to creating PDFs and because of this the common package is causing errors when
trying to run the local dev env so this change removes the healthcheck so the common package
isn't required.

The healthcheck can be re-implemented in the future once the common packages have been moved to
the common repo.