feat: add sumittion to back-end prior to submission to back-office
this has been future proofed to allow not-only appellants to submit final comments but also any
other usertype as required. The code does a scan of the database to check if any final comments
exist for a certain horizonId (note this should possibly be renamed??), if one or more exists it
checks if any of them have the same email as the one being submitted. If there is then it rejects
submission with a 409 conflict, otherwise it submits the data to the appeals beta's database. This
is is then ready for submission to horizon/the backend

#as-5788